add function to save id3 tags upon importing from soundcloud or youtube
This basic implementation solves the major problem where soundcloud or
youtube songs have no metadata on the files, so when using with a
different music player or syncing to a phone, the title will revert
to simply being the name of the file.
Still to go is a way for people with existing libraries to force an id3
tag write.