remove need to package ffmpeg binaries
To reduce binary size when potentially not needed, ffmpeg is now only
downloaded when needed on the users machine. If the binary is already
installed it will use that one otherwise it downloads the binaries to
the application data directory on the users machine.

This has been experimentally tested on OSX. I am going to test this on
Windows also before shipping a build.