Add basic boilerplate
This commit begins the work on Stretto 2.0. It includes a docker
environment, using nodejs v6 and mongodb containers.

It is also using webpack to compile js assets and node-sass to compile
scss files into css.

This all comes together in an incredibly basic front page which allows
basic theme switching through the console.

There is still much more to come, next steps are:
- React or Preact for views
- Basic view structure
- Some way of centralising all data / model methods
  so that they can be abstracted away (to be backed
  by localStorage or the server).
- Get some youtube playing!