Refactor player to pull out separate soundcloud/youtube logic
As in the title, we would like to keep these separate to keep us open
to other player options in the future. They are very tightly coupled with
their event handles right now so that is one thing that could be further
cleaned up at a later point (example of the issue this might cause is if
the youtube player fires a 'pause' event when the soundcloud player is
the currently used one, it'll change to the paused state visually but
the soundcloud player will continue to play).