Add playlist deleting and bootstrap.native dependency
Playlists can now be removed after they have been created.
An improvement to the custom Bootbox workflow is the use of
a library called `bootstrap.native` which is essentially bootstrap.js
without the dependency on jquery (hooray!). So now the modal creation
has actual animations, and the code needed to implement the modals
is simpler.