Stop soundcloud player on dispose properly
There is currently a bug where if you click a soundcloud song, but before
it loads you click a youtube song, it starts playing both songs at the same time.
To avoid this instead we load the reset url for the soundcloud player forcing
it to load an empty page that it can't play.