Implement basic syncing support
The syncing support is still very rudimentary, but it uses version numbers
between clients so that unless a change happens very minimal data needs to
be transported. However when it needs to upload it's data it has to upload
all of it (meaning the entire library is transmitted over the wire). An
improvement on this would be to instead only send the differences to the
server and for the server to manage more than just an entire 'blob' for
playlists and songs so it could make these diffs cleanly.

UX improvements could also be made by showing the status of syncing on the
sync icon and optionally on the sync view.