Add playInLibrary as an attribute to songs
I would love to be able to add all the songs my kids love listening
to to my Stretto library, however I don't want them to actually
play when I am listing to my library on shuffle. The solution to this
is to mark them as unable to play in the library and to (manually) create
a dedicated playlist where they can play just fine.